All line rules are now applied line by line
Before, we iterated over each rule and applied it to each line in a
given file.
Now, we iterate over each line and apply each rule to it.

The difference is that violations are now grouped and ordered by line
number instead of by rule, which is much more useful for users.